+++ gw2buildutil/__init__.py
@@ -1,1 +1,1 @@
-from . import api, build, defnfile, buildtemplate, compositions, textbody
+from . import api, build, buildmod, defnfile, buildtemplate, compositions, textbody